requirements.txt
Flask
googlemaps
Jinja2
simplejson
urllib3
gunicorn
connexion
firebase_admin
flask_cors
google-cloud-ndb
Pillow
google-cloud-tasks
google-python-cloud-debugger
google-cloud-storage